Extract media upload session state machine
